Docs: Update README
Based on a conversation with stakeholders from Product, NDC, and Eng, this PR seeks to update the README to provide users with all the necessary information front-loaded. Notably, we have a set of feature matrices that a connector author can use to quickly highlight what is and is not available via a particular connector. Additionally, steps to add and use the connector are included. Finally, the document is updated with various `TODO` comments to assist the author in what and how to update different aspects of the connector-specific implementations.